Rename logic package to model
