--- src/main/java/com/worldwarofants/game/logic/Wasp.java
@@ -1,6 +1,0 @@
-package com.worldwarofants.game.logic;
-
-public class Wasp extends Entity{
-
-
-}
