+++ src/main/java/com/worldwarofants/game/model/Ant.java
@@ -1,4 +1,4 @@
-package com.worldwarofants.game.logic;
+package com.worldwarofants.game.model;
 
 public class Ant extends Entity{
