+++ src/main/java/com/worldwarofants/game/model/Entity.java
@@ -1,4 +1,4 @@
-package com.worldwarofants.game.logic;
+package com.worldwarofants.game.model;
 
 /**
  * Base class to extend logic objects off of
