+++ src/main/java/com/worldwarofants/game/model/QueenAnt.java
@@ -1,6 +1,7 @@
-package com.worldwarofants.game.logic;
+package com.worldwarofants.game.model;
 
 public class QueenAnt extends Ant {
 
     private int QueenPrestige;
+
 }
