+++ src/main/java/com/worldwarofants/game/model/Wasp.java
@@ -1,0 +1,5 @@
+package com.worldwarofants.game.model;
+
+public class Wasp extends Entity{
+
+}
